% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{httpd}
\alias{httpd}
\title{Worker server}
\usage{
httpd(app, port = 9999, size = 4, preload = NULL)
}
\arguments{
\item{app}{callback function}

\item{port}{which port to run the httpd}

\item{size}{number of worker processes}

\item{preload}{character vector of packages to preload in the workers}
}
\description{
Example which handles POST / PUT requests in worker procs.
}
\examples{
\dontrun{
httpd(function(env){
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
str <- paste(capture.output(print(fm1)), collapse = "\\n")
list(
  status = 200,
  body = paste("pid:", Sys.getpid(), "\\n\\n", str),
  headers = c("Content-Type" = "text/plain")
)
}, preload = "lme4")
}
}
